% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallBearing.R
\docType{data}
\name{ballbearing}
\alias{ballbearing}
\title{Ball-bearing times}
\format{
A data set with 23 observations.
\describe{
\item{ball-bearing}{23 failure times (each measurement in 106 revolutions).}
}
}
\source{
Andrew G. Glen (2017) On the Inverse Gamma as a Survival Distribution. Journal of Quality Technology. 43(2)158-166.
}
\usage{
ballbearing
}
\description{
A data set of ball-bearing failure times.
}
\keyword{datasets}
