% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{Bspl.N}
\alias{Bspl.N}
\title{A model matrix for the B-spline}
\usage{
Bspl.N(t, knots)
}
\arguments{
\item{t}{a vector of values.}

\item{knots}{a set of knots distributed along the range of t.}
}
\description{
A model matrix for the B-spline
}
