% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{bspline}
\alias{bspline}
\title{Build the values of cubic B-spline basis function.}
\usage{
bspline(t, knots, i, m)
}
\arguments{
\item{t}{a vector of values.}

\item{knots}{a set of knots distributed along the range of t.}

\item{i}{i-th basis element in which the t values are going to be evaluated.}

\item{m}{order of the basis elements. In the case of cubic B-spline basis, m=2.}
}
\description{
Build the values of cubic B-spline basis function.
}
\references{
Carlos Alberto Cardozo Delgado, Semi-parametric generalized log-gamma regression models.  Ph. D. thesis. Sao Paulo University.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
