% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims.R
\docType{data}
\name{claims}
\alias{claims}
\title{Personal Injury Insurance Claims}
\format{
A data frame with 769 rows and 4 variables:
\describe{
\item{amount}{amount of paid money by an insurance policy or claim size, in Australian dollars}
\item{legrep}{with legal representation or not}
\item{month}{month of occurrence of the accident}
\item{optime}{operational time in percentage with range 0 to 100}
}
}
\source{
de Jong P, Heller GZ (2008) Generalized linear models for insurance data.
Cambridge University Press, Cambridge
}
\usage{
claims
}
\description{
A dataset containing personal injury insurance claims made in Australia
from January 1998 to June 1999.
}
\keyword{datasets}
