% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sglg.R
\name{coef.sglg}
\alias{coef.sglg}
\title{coef.sglg}
\usage{
\method{coef}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
coef.sglg extracts and display the estimated coefficients associated to the location parameter from a model from an object of class 'sglg'.
}
