% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{deBoor}
\alias{deBoor}
\title{Build the basis matrix and the penalty matrix  of cubic B-spline basis.}
\usage{
deBoor(x, knots)
}
\arguments{
\item{x}{a vector of values.}

\item{knots}{a set of internal knot.}
}
\description{
Build the basis matrix and the penalty matrix  of cubic B-spline basis.
}
