% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor2.R
\name{deBoor2}
\alias{deBoor2}
\title{Build the basis matrix and the penalty matrix  of cubic B-spline basis.}
\usage{
deBoor2(t, knots)
}
\arguments{
\item{t}{a vector of values.}

\item{knots}{a set of internal knot.}
}
\value{
nknot number of knots.

knots set of knots.

N  basis matrix.

K penalty matrix.
}
\description{
\code{deBoor} builds the basis matrix and penalty matrix to approximate a smooth function using
cubic B-spline cubic.
}
\examples{
t <- runif(120)
knot <- 6
deBoor2(t,knot)
}
\references{
Carlos Alberto Cardozo Delgado, Semi-parametric generalized log-gamma regression models.  Ph. D. thesis. Sao Paulo University.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
