% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.sglg.R
\name{fitted.sglg}
\alias{fitted.sglg}
\title{Extract Fitted Values}
\usage{
\method{fitted}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
fitted.sglg extracts the fitted values from a model from an object of class 'sglg'.
}
