% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.sglg.R
\name{logLik.sglg}
\alias{logLik.sglg}
\title{Extract Log-Likehood}
\usage{
\method{logLik}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
logLik.sglg extracts log-likehood from a model from an object of class 'sglg'.
}
