% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{Nbas}
\alias{Nbas}
\title{Build the basis matrix of cubic B-spline basis.}
\usage{
Nbas(x, knots)
}
\arguments{
\item{x}{a vector of values.}

\item{knots}{a set of internal knot.}
}
\description{
Build the basis matrix of cubic B-spline basis.
}
