% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sglg.R
\name{plot.sglg}
\alias{plot.sglg}
\title{plot}
\usage{
\method{plot}{sglg}(x, ...)
}
\arguments{
\item{x}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg().}

\item{...}{other arguments.}
}
\description{
plot produces the graph of the goodness-of-fit statistic which is based on the statistic Y.
Also produces a graph of the deviance-type residuals versus the fitted values for the location model.
Under the presence of a right-censored sample,
the function plot() produces a graph of the survival function of the error distribution.
}
