% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.sglg.R
\name{residuals.sglg}
\alias{residuals.sglg}
\title{Extract Model Residuals}
\usage{
\method{residuals}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
residuals.sglg extracts the deviance-type residuals for a model from an object of class 'sglg'.
}
