% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{spl.S}
\alias{spl.S}
\title{Penalized B-spline matrix, given the number of internal knots.}
\usage{
spl.S(knots)
}
\arguments{
\item{knots}{a set of internal knot.}
}
\description{
Penalized B-spline matrix, given the number of internal knots.
}
