% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sglg.R
\name{summary.sglg}
\alias{summary.sglg}
\title{summary.sglg}
\usage{
\method{summary}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
summary.sglg extracts displays the summary of the fitted model including parameter estimates, associated (approximated) standard errors and goodness-of-fit statistics from a model from an object of class 'sglg'.
}
