% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing_exercises.R
\docType{data}
\name{t_lambda}
\alias{t_lambda}
\title{######################
#                    #
# Extreme value case #
#                    #
######################}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
t_lambda
}
\description{
######################
#                    #
# Extreme value case #
#                    #
######################
}
\keyword{datasets}
