% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.sglg.R
\name{vcov.sglg}
\alias{vcov.sglg}
\title{vcov.sglg}
\usage{
\method{vcov}{sglg}(object, ...)
}
\arguments{
\item{object}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg() function.}

\item{...}{other arguments.}
}
\description{
vcov.sglg extracts the estimated variance and covariance matrix associated to parameters from a model from an object of class 'sglg'.
}
