% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBoor.R
\name{vknt}
\alias{vknt}
\title{Build a set of knots for a given number of internal knots.}
\usage{
vknt(x, knots)
}
\arguments{
\item{x}{a vector of values.}

\item{knots}{a set of internal knot.}
}
\description{
Build a set of knots for a given number of internal knots.
}
