% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.sglg.R
\name{influence.sglg}
\alias{influence.sglg}
\title{influence}
\usage{
\method{influence}{sglg}(model, ...)
}
\arguments{
\item{model}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg().}

\item{...}{other arguments.}
}
\description{
influence.sglg extracts from a object of class sglg the local influence measures and displays their graphs versus the index of the observations.
}
\references{
Carlos Alberto Cardozo Delgado, Semi-parametric generalized log-gamma regression models. Ph. D. thesis. Sao Paulo University.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
