% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss.R
\name{lss}
\alias{lss}
\title{Measures of location, scale and shape measures for a generalized log-gamma distribution}
\usage{
lss(mu, sigma, lambda)
}
\arguments{
\item{mu}{numeric, represent the location parameter of a generalized log-gamma distribution. Default value is 0.}

\item{sigma}{numeric, represent the scale parameter of a generalized log-gamma distribution. Default value is 1.}

\item{lambda}{numeric, represent the shape parameter of a generalized log-gamma distribution. Default value is 1.}
}
\description{
\code{lss} is used to obtain the mean, variance, skewness and kurtosis for a generalized log-gamma distribution.
}
\examples{
lss(0,1,-1)
lss(0,1,1)
lss(0,1,0.005)
}
\references{
Carlos Alberto Cardozo Delgado, Semi-parametric generalized log-gamma regression models. Ph. D. thesis. Sao Paulo University.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
