% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.npc.R
\name{plotnpc}
\alias{plotnpc}
\title{Plotting a natural cubic splines or P-splines.}
\usage{
plotnpc(fit)
}
\arguments{
\item{fit}{an object of the class sglg. This object is returned from the call to glg(), sglg(), survglg() or ssurvglg().}
}
\description{
plotnpc displays a graph of a fitted nonparametric effect, either natural cubic spline or P-spline, from an object of class sglg.
}
\references{
Eilers P.H.C. and Marx B.D. (1996). Flexible smoothing with B-splines and penalties. Statistical Science. 11, 89-121.

Wood, S. (2006). Additive generalized models: An R introduction. Chapman and Hall.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
