% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglg2.R
\name{sglg}
\alias{sglg}
\title{Fitting semi-parametric generalized log-gamma regression models}
\usage{
sglg(
  formula,
  npc,
  basis,
  data,
  shape,
  method,
  alpha0,
  nknts,
  Tolerance,
  Maxiter
)
}
\arguments{
\item{formula}{a symbolic description of the systematic component of the model to be fitted. See details for further information.}

\item{npc}{a data frame with potential nonparametric variables of the systematic part of the model to be fitted.}

\item{basis}{a name of the cubic spline basis to be used in the model. Supported basis include deBoor and Gu basis
which are a B-spline basis and a natural cubic spline basis, respectively.}

\item{data}{an optional data frame, list containing the variables in the model.}

\item{shape}{an optional value for the shape parameter of the error distribution of a generalized log-gamma distribution. Default value is 1.}

\item{method}{There are two possibles algorithms to estimate the parameters. The default algorithm is 'FS' Fisher-Scoring,
the other option is 'GSFS' an adequate combination between the block matrix version of non-linear Gauss-Seidel algorithm and Fisher-Scoring algorithm.}

\item{alpha0}{is a vector of initial values for the smoothing parameter alpha.}

\item{nknts}{is a vector of the number of knots in each non-linear component of the model.}

\item{Tolerance}{an optional positive value, which represents the convergence criterion. Default value is 1e-04.}

\item{Maxiter}{an optional positive integer giving the maximal number of iterations for the estimating process. Default value is 1e03.}
}
\value{
mu a vector of parameter estimates asociated with the location parameter.

sigma estimate of the scale parameter associated with the model.

lambda estimate of the shape parameter associated with the model.

interval estimate of a 95\% confidence interval for each estimate parameters associated with the model.

Deviance the deviance associated with the model.
}
\description{
\code{sglg} is used to fit a semi-parametric regression model suitable for analysis of data sets in which the response variable is continuous, strictly positive, and asymmetric.
In this setup, the location parameter of the response variable is explicitly modeled by semi-parametric functions, whose nonparametric components may be approximated by
natural cubic splines or cubic P-splines.
}
\examples{
library(sglg)
set.seed(1)
n <- 300
error <- rglg(n,0,1,1)
x1 <- runif(n,-3,3)
x2 <- rbinom(n,1,0.5)
t <- as.matrix((2*1:n - 1)/(2*n))
colnames(t) <- "t"
f_t <- cos(4*pi*t)
y <- 2*x1 + 0.5*x2 + f_t + error
colnames(y) <- "y"
data <- as.data.frame(cbind(y,x1,x2,t))
fit1 <- sglg(y ~ x1 + x2 - 1,npc=t,data=data,basis = "deBoor")
logLik(fit1)
fit2 <- sglg(y ~ x1 + x2 - 1,npc=t,data=data,basis = "Gu",alpha0=c(0.05,0.1),nknts=7)
logLik(fit2)
}
\references{
Carlos A. Cardozo, G. Paula and L. Vanegas. Semi-parametric generalized log-gamma regression models. In preparation.
}
\author{
Carlos Alberto Cardozo Delgado <cardozorpackages@gmail.com>, G. Paula and L. Vanegas.
}
